drivers: interrupt_controller: loapic: address MISRA Rule 7.2 and 10.6
Make the LOAPIC code conform to MISRA C:2012:

  - Rule 7.2: add the 'U' suffix to the 0xffffffff DFR write constant,
    whose value is representable only in an unsigned type.
  - Rule 10.6: cast apic_id to uint32_t before the 24-bit left shift so
    the composite expression has the same essential type as the 32-bit
    ICRHI register argument.

No functional change.

+++ include/zephyr/drivers/interrupt_controller/loapic.h
@@ -173,7 +173,7 @@ static inline void z_loapic_ipi(uint8_t apic_id, uint32_t ipi, uint8_t vector)
 	while (x86_read_xapic(LOAPIC_ICRLO) & LOAPIC_ICR_BUSY) {
 	}
 
-	x86_write_xapic(LOAPIC_ICRHI, apic_id << 24);
+	x86_write_xapic(LOAPIC_ICRHI, (uint32_t)apic_id << 24);
 	x86_write_xapic(LOAPIC_ICRLO, ipi);
 #else
 	/*
